+++ src/OrcAI.Local/AssemblyInfo.fs
@@ -1,0 +1,6 @@
+module OrcAI.Local.AssemblyInfo
+
+open System.Runtime.CompilerServices
+
+[<assembly: InternalsVisibleTo("OrcAI.Local.Tests")>]
+do ()